+++ src/custom-types/replace-property.ts
@@ -1,0 +1,5 @@
+type ReplaceProperty<Type, ReplaceKey extends keyof Type, NewKey> = Omit<Type, ReplaceKey> & {
+  [Property in ReplaceKey]: NewKey
+};
+
+export { ReplaceProperty };